LayoutTests/media/track/track-webvtt-tc028-unsupported-markup-expected.txt

Tests that unsupported markup is properly ignored.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'Bear is Coming!!!!! And what kind of a bear it is - just have look.') OK
EXPECTED (cues[i].getCueAsSource() == '
Bear is Coming!!!!!

And what kind of a bear it is - just have look.

') OK
EXPECTED (cues[i].getCueAsHTML().textContent == ' I said Bear is coming!!!! I said Bear is still coming!!!! ') OK
EXPECTED (cues[i].getCueAsSource() == '
I said Bear is coming!!!!
I said Bear is still coming!!!!
') OK
EXPECTED (cues[i].getCueAsHTML().textContent == ' I said Bear is coming now!!!! ') OK
EXPECTED (cues[i].getCueAsSource() == '
I said Bear is coming now!!!!

') OK
END OF TEST
